ScriptFileTransferEmby/requirements.txt
requests
watchdog
python-dotenv
